Add pundit policy to all (not_admin) controllers

+++ app/controllers/accounts_controller.rb
@@ -3,19 +3,26 @@ class AccountsController < BaseController
   before_action :set_account, only: %i[show edit update destroy]
 
   def index
-    @accounts = current_user.accounts
+    index_accounts
+    authorize index_accounts
   end
 
-  def show; end
+  def show
+    authorize set_account
+  end
 
   def new
-    @account = Account.new
+    new_account
+    authorize new_account
   end
 
-  def edit; end
+  def edit
+    authorize set_account
+  end
 
   def create
-    if new_account.save
+    authorize create_account
+    if create_account.save
       redirect_to root_path, notice: 'Account was successfully created.'
     else
       render :new
@@ -23,6 +30,7 @@ class AccountsController < BaseController
   end
 
   def update
+    authorize set_account
     if set_account.update(account_params)
       redirect_to root_path, notice: 'Account was successfully updated.'
     else
@@ -31,6 +39,7 @@ class AccountsController < BaseController
   end
 
   def destroy
+    authorize set_account
     set_account.destroy
     redirect_to root_path, notice: 'Account was successfully destroyed.'
   end
@@ -42,7 +51,15 @@ class AccountsController < BaseController
   end
 
   def new_account
+    @account = Account.new
+  end
+
+  def create_account
     @account = Account.new(account_params)
+  end
+
+  def index_accounts
+    @accounts = current_user.accounts
   end
 
   def account_params
